# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drone.Works API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "def-0": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pointer",
          "message"
        ],
        "properties": {
          "pointer": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "title": "ProblemError"
      },
...
    }
  },
  "paths": {
...
    "/api/v1/organizations/{organization_id}/invitations": {
      "post": {
        "operationId": "createInvitation",
        "summary": "Create an organization invitation",
        "tags": [
          "invitations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "email",
                  "role"
                ],
                "properties": {
                  "email": {
                    "minLength": 3,
                    "maxLength": 320,
                    "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
                    "type": "string"
                  },
                  "role": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "admin"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "pilot"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "viewer"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "organization_id",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "A single-use organization invitation was created.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "invitation_id",
                    "organization_id",
                    "role",
                    "expires_at"
                  ],
                  "properties": {
                    "invitation_id": {
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                      "type": "string"
                    },
                    "organization_id": {
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                      "type": "string"
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "admin"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "pilot"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "viewer"
                          ]
                        }
                      ]
                    },
                    "expires_at": {
                      "format": "date-time",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "The request was denied or could not be accepted.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          },
          "5XX": {
            "description": "The request could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/invitations/accept": {
      "post": {
        "operationId": "acceptInvitation",
        "summary": "Accept an organization invitation",
        "tags": [
          "invitations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "token"
                ],
                "properties": {
                  "token": {
                    "minLength": 43,
                    "maxLength": 128,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "organization_id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The verified matching user accepted the invitation.",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "user_id",
                    "role",
                    "pilot_profile_id"
                  ],
                  "properties": {
                    "user_id": {
                      "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                      "type": "string"
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "owner"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "admin"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "pilot"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "viewer"
                          ]
                        }
                      ]
                    },
                    "pilot_profile_id": {
                      "anyOf": [
                        {
                          "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "4XX": {
            "description": "The request was denied or could not be accepted.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          },
          "5XX": {
            "description": "The request could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/invitations/{invitation_id}": {
      "delete": {
        "operationId": "revokeInvitation",
        "summary": "Revoke an organization invitation",
        "tags": [
          "invitations"
        ],
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "organization_id",
            "required": true
          },
          {
            "schema": {
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "type": "string"
            },
            "in": "path",
            "name": "invitation_id",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "The pending invitation was revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The pending invitation was revoked."
                }
              }
            }
          },
          "4XX": {
            "description": "The request was denied or could not be accepted.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          },
          "5XX": {
            "description": "The request could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail",
                    "instance",
                    "correlation_id"
                  ],
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "minimum": 400,
                      "maximum": 599,
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "correlation_id": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}